feat: add package utils
