+++ packages/keychain/package.json
@@ -1,4 +1,5 @@
 {
+<<<<<<< HEAD
 	"name": "@ts-drp/keychain",
 	"version": "0.8.5",
 	"license": "MIT",
@@ -38,4 +39,97 @@
 	"devDependencies": {
 		"@types/object-inspect": "^1.13.0"
 	}
+||||||| parent of 71158af (feat: add package utils)
+  "name": "@ts-drp/keychain",
+  "version": "0.8.5",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/topology-foundation/ts-drp.git"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
+  "scripts": {
+    "build": "tsc -b",
+    "clean": "rm -rf dist/ node_modules/",
+    "prepack": "tsc -b",
+    "watch": "tsc -b -w"
+  },
+  "dependencies": {
+    "@bufbuild/protobuf": "^2.0.0",
+    "@chainsafe/bls": "^8.1.0",
+    "@chainsafe/bls-keygen": "^0.4.0",
+    "@libp2p/crypto": "^5.0.5",
+    "@libp2p/interface": "^2.1.3",
+    "@ts-drp/object": "0.8.5",
+    "uint8arrays": "^5.1.0"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.18.0",
+    "@types/object-inspect": "^1.13.0",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
+    "eslint": "^9.18.0",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.20.0"
+  }
+=======
+  "name": "@ts-drp/keychain",
+  "version": "0.8.5",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/drp-tech/ts-drp.git"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
+  "scripts": {
+    "build": "tsc -b",
+    "clean": "rm -rf dist/ node_modules/",
+    "prepack": "tsc -b",
+    "watch": "tsc -b -w"
+  },
+  "dependencies": {
+    "@bufbuild/protobuf": "^2.0.0",
+    "@chainsafe/bls": "^8.1.0",
+    "@chainsafe/bls-keygen": "^0.4.0",
+    "@libp2p/crypto": "^5.0.5",
+    "@libp2p/interface": "^2.1.3",
+    "@ts-drp/object": "0.8.5",
+    "uint8arrays": "^5.1.0"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.18.0",
+    "@types/object-inspect": "^1.13.0",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
+    "eslint": "^9.18.0",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.20.0"
+  }
+>>>>>>> 71158af (feat: add package utils)
 }
