+++ packages/keychain/src/keychain.ts
@@ -66,3 +66,4 @@ export class Keychain {
 		return this._ed25519PrivateKey.raw;
 	}
 }
+
