+++ packages/types/src/index.ts
@@ -22,3 +22,4 @@ export {
 } from "./proto/drp/v1/messages_pb.js";
 
 export * from "./hashgraph.js";
+export * from "./interval-runner.js";
